+++ public_html/index.php
@@ -35,7 +35,7 @@ require __DIR__.'/../bootstrap/autoload.php';
 $app = require_once __DIR__.'/../bootstrap/app.php';
 
 // set the public path to this directory
-$app->bind('path.public', function() {
+$app->bind('path.public', function () {
     return __DIR__;
 });